Redesign ir::Node: merge control/data dependencies.
Separate types of dependencies are replaced by support of multiple return value.
This is the first of a series commits for this refactoring.
This commit reverts 4959d79, 856c688, 58329df.
The rational behind this change is that multiple return values could be more useful for instruction selection and M-extension (e.g. div/rem).